+++ pennylane/ops/qubit/__init__.py
@@ -35,6 +35,7 @@ from .observables import *
 from .parametric_ops import *
 from .qchem_ops import *
 from .state_preparation import *
+from .hamiltonian import Hamiltonian
 
 ops = {
     "Hadamard",
@@ -96,6 +97,7 @@ obs = {
     "Hermitian",
     "Projector",
     "SparseHamiltonian",
+    "Hamiltonian"
 }
