+++ tests/test_operation.py
@@ -128,6 +128,9 @@ class TestOperation:
         if test_class == qml.QubitUnitary:
             pytest.skip("QubitUnitary can act on any number of wires.")
 
+        if test_class == qml.Hamiltonian:
+            pytest.skip("Hamiltonian has a different initialization signature.")
+
         if test_class in (qml.ControlledQubitUnitary, qml.MultiControlledX):
             pytest.skip("ControlledQubitUnitary alters the input params and wires in its __init__")
